separate types module
